separate sources to 'ui' and 'core'. Fill JavaDoc a little bit
