+++ CMakeLists.txt
@@ -5,10 +5,8 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 set(CMAKE_AUTOMOC ON)
-set(CMAKE_AUTORCC ON)
-set(CMAKE_AUTOUIC ON)
 
-find_package(Qt5
+find_package(Qt5 QUIET
         COMPONENTS Widgets)
 if (Qt5_FOUND)
     set(QtVersion 5)
@@ -19,11 +17,18 @@ else ()
 endif ()
 
 add_executable(mdflash
-    main.cpp
-    markdown_processor.cpp
-)
+        main.cpp)
 
-target_link_libraries(mdflash PRIVATE Qt${QtVersion}::Widgets)
+file(GLOB_RECURSE Sources src/*.cpp)
+
+target_sources(mdflash
+        PRIVATE ${Sources})
+
+target_link_libraries(mdflash
+        PRIVATE Qt${QtVersion}::Widgets)
+
+target_include_directories(mdflash
+        PUBLIC include
+        PRIVATE include/mdflash)
 
 install(TARGETS mdflash DESTINATION bin)
-
